ageDescription solved timeForHalfWay solved whichRookThreatens solved rookOrBishopThreatens solved triangleKind solved All ifElse solved Fixed ageDescription and timeForHalfWay isNumberHappy solved All Logical solved ageDescription fixed lesson3 to isCoPrime solved squareBetweenExists solved collatzSteps solved sin solved sin changed squareSequenceDigit solved without fib Long (Let it be Int)

+++ src/lesson3/task1/Loop.kt
@@ -93,7 +93,7 @@ fun digitNumber(n: Int): Int {
  * Найти число Фибоначчи из ряда 1, 1, 2, 3, 5, 8, 13, 21, ... с номером n.
  * Ряд Фибоначчи определён следующим образом: fib(1) = 1, fib(2) = 1, fib(n+2) = fib(n) + fib(n+1)
  */
-fun fib(n: Long): Long {
+fun fib(n: Int): Int {
     var nm2: Long = 1
     var nm1: Long = 1
     var nn: Long = 0
@@ -103,7 +103,7 @@ fun fib(n: Long): Long {
             nm2 = nm1
             nm1 = nn
         }
-        nn
+        nn.toInt()
     }
 }
